Change indentation to match other files (#1034)
